+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - debian: dbconfig allow major upgrade without backup dump [PR #2538]
 - plugins: mariabackup python-mysqlclient ignore invalid utf8 [PR #2560]
 - dedup backend: Fix leaking filedescriptor [PR #2549]
+- vol-mgr: fix not locking the volume reservation chain during copying [PR #2582]
 
 ### Removed
 - dird: deprecate Pool->FileRetention, Pool->JobRetention, WriteVerifyList [PR #2573]
@@ -2237,4 +2238,5 @@ If you want to migrate from your manually configured disk autochanger to simply 
 [PR #2568]: https://github.com/bareos/bareos/pull/2568
 [PR #2573]: https://github.com/bareos/bareos/pull/2573
 [PR #2574]: https://github.com/bareos/bareos/pull/2574
+[PR #2582]: https://github.com/bareos/bareos/pull/2582
 [unreleased]: https://github.com/bareos/bareos/tree/master